+++ backend/src/presentation/controllers/gift/create-gift.ts
@@ -1,7 +1,7 @@
-import { CreateGift } from "../../domain/useCases/gift"
-import { HttpRequest, HttpResponse, ok, serverError } from "../contracts"
-import { Controller } from "../contracts/controller"
-import { GiftViewModel } from "../view-models"
+import { CreateGift } from "../../../domain/useCases/gift"
+import { HttpRequest, HttpResponse, ok, serverError } from "../../contracts"
+import { Controller } from "../../contracts/controller"
+import { GiftViewModel } from "../../view-models"
 
 export class CreateGiftController implements Controller {
   constructor(private readonly createGiftService: CreateGift) { }
