+++ backend/src/presentation/controllers/gift/get-gift-by-id.ts
@@ -1,0 +1,18 @@
+import { GetGiftById } from "../../../domain/useCases/gift"
+import { HttpRequest, HttpResponse, ok, resourceNotFoundError, serverError } from "../../contracts"
+import { Controller } from "../../contracts/controller"
+import { GiftViewModel } from "../../view-models"
+
+export class GetGiftByIdController implements Controller {
+  constructor(private readonly getGiftByIdService: GetGiftById) { }
+
+  async handle(req?: HttpRequest<{ giftId: string }>): Promise<HttpResponse<GiftViewModel>> {
+    try {
+      const foundGift = await this.getGiftByIdService.execute(req?.body?.giftId)
+      if (!foundGift) return resourceNotFoundError('gift')
+      return ok(foundGift)
+    } catch (error) {
+      return serverError(error as Error)
+    }
+  }
+}
