+++ backend/src/presentation/controllers/gift/index.ts
@@ -1,1 +1,3 @@
 export * from './create-gift'
+export * from './delete-gift'
+export * from './get-gift-by-id'
